interpretation/requirements.txt
numpy>=1.17.0
azure-cli-core
azureml-sdk==1.0.74